refactor: move utils folder into single file on the lib folder
