+++ src/services/posts.ts
@@ -1,7 +1,6 @@
+import { pick } from 'contentlayer/client'
 import type { Post } from 'contentlayer/generated'
 import { allPosts } from 'contentlayer/generated'
-
-import { pick } from '@/utils'
 
 export const getAllPosts = (fields: (keyof Post)[] = []): Array<Post> => {
   const filteredPosts = allPosts
