+++ src/services/snippets.ts
@@ -1,7 +1,6 @@
+import { pick } from 'contentlayer/client'
 import type { Snippet } from 'contentlayer/generated'
 import { allSnippets } from 'contentlayer/generated'
-
-import { pick } from '@/utils'
 
 export const getAllSnippets = (
   fields: (keyof Snippet)[] = [],
