--- src/utils/array.ts
@@ -1,11 +1,0 @@
-export const unique = <T, Key extends keyof T>(
-  array: Array<T> | T[],
-  property?: Key,
-): Array<T> => {
-  if (!property) return Array.from(new Set([...array]))
-
-  const set = new Set()
-  return array.filter((o: T) => {
-    return !set.has(o[property]) && set.add(o[property])
-  })
-}
