--- src/utils/index.ts
@@ -1,4 +1,0 @@
-export * from './array'
-export * from './object'
-export * from './string'
-export * from './transformers'
