--- src/utils/object.ts
@@ -1,11 +1,0 @@
-import type { ConvertPick } from '@/types'
-
-export const pick = <Obj, Keys extends keyof Obj>(
-  obj: Obj,
-  keys: Keys[],
-): ConvertPick<{ [K in Keys]: Obj[K] }> => {
-  return keys.reduce((acc, key) => {
-    acc[key] = obj[key] || null
-    return acc
-  }, {} as any)
-}
